% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratXtrait_fit.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife estimates}
\usage{
jackknife(z1, z2, r, cat_mems, N1, N2, N_refpop, blocks = 20,
  weighted = FALSE, bias_correction = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Jackknife estimates
}

