% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratXtrait_draws_3.R
\name{rand_sim_data}
\alias{rand_sim_data}
\title{Generate random simulation parameters to pass to sim1}
\usage{
rand_sim_data(N_snp, N1, N2 = NULL, Ns, N_refpop, .cov = NULL,
  cat_sizes = NULL, cat_props = c(1, 0.25, 0.4), cat_mems = NULL,
  cat_mats = NULL, shrink = TRUE)
}
\arguments{
\item{shrink}{}
}
\description{
Generate random simulation parameters to pass to sim1
}

